Bug fixing in the loader
When you had gregorio.tex and gregorio.lua, the loader tried to
load gregoriotex.tex as a lua file, and obviously failed.

I added a check on the extention in the case where we look for the
name without any, and also reorganized the search order to avoid
looking for xxx.texluac before looking for xxx.lua.